+++ trilby-cli/src/Trilby/Infect.hs
@@ -5,7 +5,7 @@ import Data.Text qualified as Text
 import Data.Text.IO qualified as Text
 import Trilby.Configuration (Configuration (..))
 import Trilby.Configuration qualified as Configuration
-import Trilby.HNix (FileOrFlake (..), nixBuild, trilbyFlake, writeNixFile)
+import Trilby.HNix (FileOrFlake (..), copyClosure, nixBuild, trilbyFlake, writeNixFile)
 import Trilby.Host
 import Trilby.Infect.Options
 import Prelude
@@ -15,13 +15,12 @@ infect (askOpts -> opts) = do
     configurations <- mapM Configuration.fromHost . NonEmpty.nubOrd =<< opts.hosts
     kexec <- buildKexec opts
     for_ configurations $ \Configuration{..} -> do
-        targetPath <- extractKexec host kexec
-        let binPath = targetPath </> $(mkRelDir "kexec_trilby/bin/kexec-trilby")
-        whenM opts.reboot $ ssh host rawCmd_ [fromPath binPath]
+        copyClosure host kexec
+        let bin = kexec </> $(mkRelFile "kexec-boot")
+        whenM opts.reboot $ ssh host rawCmd_ ["sudo", fromPath bin]
 
-buildKexec :: (HasCallStack) => InfectOpts App -> App (Path Abs File)
-buildKexec opts = withSystemTempFile "trilby-infect-.nix" $ \tmpFile handle -> do
-    hClose handle
+buildKexec :: (HasCallStack) => InfectOpts App -> App (Path Abs Dir)
+buildKexec opts = withTempFile $(mkRelFile "infect.nix") \tmpFile -> do
     trilbyUrl <- trilbyFlake
     edition <- opts.edition
     authorisedKeys <-
@@ -49,23 +48,4 @@ buildKexec opts = withSystemTempFile "trilby-infect-.nix" $ \tmpFile handle -> d
         };
         in system.config.system.build.kexecTree
         |]
-    resultDir <- nixBuild $ File (Abs tmpFile)
-    pure $ resultDir </> $(mkRelFile "tarball/trilby-kexec.tar")
-
-extractKexec :: (HasCallStack) => Host -> Path b File -> App (Path Abs Dir)
-extractKexec host srcPath = do
-    let dstDir = $(mkAbsDir "/")
-    case host of
-        Localhost -> cmd_ ["tar", "-xf", fromPath srcPath, "-C", fromPath dstDir]
-        Host{} ->
-            let command =
-                    Text.unwords . mconcat $
-                        [ ["cat", fromPath srcPath]
-                        , ["|"]
-                        , ["ssh", ishow host]
-                        , ["'"]
-                        , ["tar", "-xf", "-", "-C", fromPath dstDir]
-                        , ["'"]
-                        ]
-             in void $ shell command empty
-    pure dstDir
+    nixBuild (File $ Abs tmpFile) parseAbsDir